agent-runner-service/pyproject.toml
[project]
name = "orchestrator-service"
version = "0.1.0"
description = "Orchestrator service that coordinates persona testing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "streamlit>=1.28.0",
    "httpx>=0.25.2",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "black>=25.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
